# src/main/resources/docs/api.yaml
openapi: 3.1.0
info:
  title: HFF Data Studio API
  description: HFF Data Intelligence Platform - CATalyst API
  version: 1.0.0
  contact:
    name: Data Analyst Service - HFF
    email: data-analyst@hff.re
  license:
    name: HFF Internal Use - Restricted Distribution
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://catalyst.hff.re/api/v1
    description: Production server
...
paths:
...
  /files/upload:
    post:
      tags: [Files]
      summary: Upload a file
      operationId: uploadFile
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: fileType
          in: query
          schema:
            $ref: "#/components/schemas/FileType"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceFileResponse"
  /files/{fileId}:
    get:
      tags: [Files]
      summary: Get file by ID
      operationId: getFileById
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: File details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceFileResponse"
    delete:
      tags: [Files]
      summary: Delete file
      operationId: deleteFile
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: File deleted
  /files/project/{projectId}:
    get:
      tags: [Files]
      summary: Get files by project
      operationId: getFilesByProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceFileResponse"
  /files/project/{projectId}/page:
    get:
      tags: [Files]
      summary: Get files by project (paginated)
      operationId: getFilesByProjectPaged
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageSourceFileResponse"
  /files/{fileId}/datasets:
    get:
      tags: [Files]
      summary: Get datasets by file
      operationId: getDatasetsByFile
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DatasetResponse"
  /files/{fileId}/download:
    get:
      tags: [Files]
      summary: Download file
      operationId: downloadFile
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /files/{fileId}/preview:
    get:
      tags: [Files]
      summary: Get file preview
      operationId: getFilePreview
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: maxLines
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: File preview content
          content:
            text/plain:
              schema:
                type: string
  /files/{fileId}/statistics:
    get:
      tags: [Files]
      summary: Get file statistics
      operationId: getFileStatistics
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: File statistics
          content:
            application/json:
              schema:
                type: object
  /files/status/{status}:
    get:
      tags: [Files]
      summary: Get files by status
      operationId: getFilesByStatus
      parameters:
        - name: status
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FileProcessingStatus"
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceFileResponse"
  /files/{fileId}/status:
    patch:
      tags: [Files]
      summary: Update file status
      operationId: updateFileStatus
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/FileProcessingStatus"
      responses:
        "200":
          description: File status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceFileResponse"
  /files/all:
    get:
      tags: [Files]
      summary: Get all files (paginated)
      operationId: getAllFiles
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageSourceFileResponse"

  # ==================== DATASETS ====================
...
  /datasets/file/{fileId}:
    get:
      tags: [Datasets]
      summary: Get datasets by file
      operationId: getDatasetsByFile
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of datasets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageDatasetResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FileType:
      type: string
      enum: [CSV, EXCEL, SQL]
    FileProcessingStatus:
      type: string
      enum: [RECEIVED, ANALYZING, EXPLORED, CLEANED, ERROR]
    SourceFileResponse:
      type: object
      properties:
        fileId:
          type: string
          format: uuid
        fileName:
          type: string
        fileType:
          $ref: "#/components/schemas/FileType"
        storagePath:
          type: string
        sizeBytes:
          type: integer
          format: int64
        sizeFormatted:
          type: string
        uploadedAt:
          type: string
          format: date-time
        processingStatus:
          $ref: "#/components/schemas/FileProcessingStatus"
        statusLabel:
          type: string
        projectId:
          type: string
          format: uuid
        projectName:
          type: string
        userId:
          type: string
          format: uuid
        userFullName:
          type: string
        datasetCount:
          type: integer

    # ===== DATASETS =====
    DatasetResponse:
      type: object
      properties:
        datasetId:
          type: string
          format: uuid
        datasetName:
          type: string
        rowCount:
          type: integer
        columnCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        isCleaned:
          type: boolean
        fileId:
          type: string
          format: uuid
        fileName:
          type: string
        projectId:
          type: string
          format: uuid
        projectName:
          type: string
        qualityScore:
          type: integer
        analysisCount:
          type: integer
...
    PageResponse:
      type: object
      properties:
        content:
          type: array
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        isFirst:
          type: boolean
        isLast:
          type: boolean
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
    PageSourceFileResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/SourceFileResponse"
    PageDatasetResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/DatasetResponse"
...
security:
  - bearerAuth: []